lc: implement per client formatting
So here's the deal: I'm using two language servers for Python, and only
one of them can format the code. So instead of relying on buf_request, I
need to send the request to the client. This implementation is generic
in which it will send the request to the one registered most recently.

This also changes the logic so we only setup the <leader>f mapping if
the server actually supports it.