No more fun
The library was kinda ok, but buggy, and it's abandonware at this point.
The code was nicer, but I don't want to suffer maintaining. I did
introduce tests to the repo as part adding fun, so I'll keep that (with
a dummy test for now).

Work for tomorrow: add some tests, with mocking for the vim global (or
maybe run tests with nvim --headless, I wanted to avoid that as much as
possible, but it's cool to have shit like vim.inspect in tests).
Alternatively, I can figure out how to bring vim utilities to LUA_PATH?