Initial support for textDocument/codeLens
This first version supports fetching and rendering codeLens in virtual
text. Ideally, the codeLens would be refreshed as a reaction to
textDocument/didChange, but currently neovim doesn't support plugging
into that, so this version of the code leverages as side-effect of
didChange: the call to publishDiagnostics (if we happen to be unlucky
with a server that supports codelens but doesn't support diagnostics, I
can think of a better solution). I'll raise an issue in the neovim repo
where we can discuss this.

Some other changes here, include the change to the publishDiagnostics
handler configuration: virtual text is now going to be used for
codelens, so let's not render diagnostics in the virtual text. Moreover,
in order to make my life easier with diagnostics, I changed the mapping
for show_line_diagnostics to be <leader><leader>, since I am losing the
promptly available information about diagnostics. As a cleanup, I've
also removed the highlight customizations and the overriding of
get_virtual_text_chunks_for_line.

Tomorrow I'll work on codeLens/resolve. There's also some opportunities
for improving re-rendering (for example, in the current version the
virtual text "blinks" every time, but we could be smarter about which
entries to clear, doing some form of diff on the code lens).

Implementing codeLens/resolve will require caching the results of
textoDocument/codeLens, so we can also reuse that cache for improving
rendering.