Migrate from diagnosticls to efm-langserver
efm-langserver has improved on the issues I used to have with it.
Namely, not supporting LSP settings and not supporting debounce.
diagnosticls still supports per tool debouncing, but I had 250ms for all
tools, which is supported by efm.

One last thing that efm doesn't have yet, is support for specifying the
source for diagnostic, but I sent a PR for that (see
mattn/efm-langserver#80). The code here is already using the parameter,
even though it won't work for now (if I get annoyed by it I can install
efm-langserver from my fork, but it's not a big deal).