Add vim-wordmotion
So this make vim behave like emacs, please don't hate me. I actually
don't know if I'm going to get used to it, but let me give it a try.
Worst case, I can setup a leader for the motions, but I don't know if I
want to do that.

Using this instead of the builtin settings because I still want <cword>
to behave "normally".